1. Add a new riscv platform: axu15eg 2. Rename the riscv64-qemu-virt platfrom to riscv64-common
