+++ modules/axhal/src/platform/mod.rs
@@ -11,8 +11,11 @@ cfg_if::cfg_if! {
         mod x86_pc;
         pub use self::x86_pc::*;
     } else if #[cfg(all(target_arch = "riscv64", platform_family = "riscv64-qemu-virt"))] {
-        mod riscv64_qemu_virt;
-        pub use self::riscv64_qemu_virt::*;
+        mod riscv64_common;
+        pub use self::riscv64_common::*;
+    } else if #[cfg(all(target_arch = "riscv64", platform_family = "riscv64-axu15eg"))] {
+        mod riscv64_common;
+        pub use self::riscv64_common::*;
     } else if #[cfg(all(target_arch = "aarch64", platform_family = "aarch64-qemu-virt"))] {
         mod aarch64_qemu_virt;
         pub use self::aarch64_qemu_virt::*;
