+++ modules/axhal/src/platform/riscv64_common/console.rs
