+++ modules/axhal/src/platform/riscv64_common/misc.rs
