+++ platforms/riscv64-axu15eg.toml
@@ -1,0 +1,29 @@
+# Architecture identifier.
+arch = "riscv64"
+# Platform identifier.
+platform = "riscv64-axu15eg"
+# Platform family.
+family = "riscv64-axu15eg"
+
+# Base address of the whole physical memory.
+phys-memory-base = "0x8000_0000"
+# Size of the whole physical memory.
+phys-memory-size = "0x800_0000"     # 128M
+# Base physical address of the kernel image.
+kernel-base-paddr = "0x8020_0000"
+# Base virtual address of the kernel image.
+kernel-base-vaddr = "0xffff_ffc0_8020_0000"
+# Linear mapping offset, for quick conversions between physical and virtual
+# addresses.
+phys-virt-offset = "0xffff_ffc0_0000_0000"
+# MMIO regions with format (`base_paddr`, `size`).
+mmio-regions = [
+    ["0x0c00_0000", "0x400_0000"],   # PLIC
+    ["0x1000_0000", "0x100_0000"],   # ATSINTC
+    ["0x6000_0000", "0x2000_0000"],  # MMIO PORT
+]
+# VirtIO MMIO regions with format (`base_paddr`, `size`).
+virtio-mmio-regions = []
+
+# Timer interrupt frequency in Hz.
+timer-frequency = "10_000_000"      # 10MHz
